Consolidate Controller/Service/Repository code Fixes #161
