+++ api/src/main/java/org/endeavourhealth/imapi/controllers/ConceptController.java
@@ -2,10 +2,8 @@ package org.endeavourhealth.imapi.controllers;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.endeavourhealth.imapi.config.ConfigManager;
+import org.endeavourhealth.imapi.logic.service.ConceptService;
 import org.endeavourhealth.imapi.logic.service.EntityService;
-import org.endeavourhealth.imapi.logic.service.RequestObjectService;
-import org.endeavourhealth.imapi.logic.service.SetService;
 import org.endeavourhealth.imapi.model.ConceptContextMap;
 import org.endeavourhealth.imapi.model.EntityReferenceNode;
 import org.endeavourhealth.imapi.model.Pageable;
@@ -31,13 +29,13 @@ import java.util.*;
 public class ConceptController {
 
   private static final Logger LOG = LoggerFactory.getLogger(ConceptController.class);
-  private final EntityService entityService = new EntityService();
+  private final ConceptService conceptService = new ConceptService();
 
   @GetMapping(value = "/public/matchedFrom", produces = "application/json")
   public Collection<SimpleMap> getMatchedFrom(@RequestParam(name = "iri") String iri) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.MatchedFrom.GET")) {
       LOG.debug("getMatchedFrom");
-      return entityService.getMatchedFrom(iri);
+      return conceptService.getMatchedFrom(iri);
     }
   }
 
@@ -45,7 +43,7 @@ public class ConceptController {
   public Collection<SimpleMap> getMatchedTo(@RequestParam(name = "iri") String iri) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.MatchedTo.GET")) {
       LOG.debug("getMatchedTo");
-      return entityService.getMatchedTo(iri);
+      return conceptService.getMatchedTo(iri);
     }
   }
 
@@ -53,70 +51,43 @@ public class ConceptController {
   public List<SearchTermCode> getTermCodes(@RequestParam(name = "iri") String iri, @RequestParam(name = "includeInactive") Optional<Boolean> includeInactive) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.TermCode.GET")) {
       LOG.debug("getTermCodes");
-      return entityService.getEntityTermCodes(iri, includeInactive.orElseGet(() -> false));
+      return conceptService.getEntityTermCodes(iri, includeInactive.orElseGet(() -> false));
     }
   }
 
   @GetMapping(value = "/public/superiorPropertiesPaged")
-  @Operation(
-    summary = "Get top level properties for an entity as a tree node",
-    description = "Finds the highest parent (superior) properties for an entity and returns then in a tree node format for use in a hierarchy tree"
-  )
-  public Pageable<EntityReferenceNode> getSuperiorPropertiesPaged(
-    @RequestParam(name = "conceptIri") String iri,
-    @RequestParam(name = "schemeIris", required = false) List<String> schemeIris,
-    @RequestParam(name = "page", required = false) Integer page,
-    @RequestParam(name = "size", required = false) Integer size,
-    @RequestParam(name = "inactive", required = false) boolean inactive
-  ) throws IOException {
+  @Operation(summary = "Get top level properties for an entity as a tree node", description = "Finds the highest parent (superior) properties for an entity and returns then in a tree node format for use in a hierarchy tree")
+  public Pageable<EntityReferenceNode> getSuperiorPropertiesPaged(@RequestParam(name = "conceptIri") String iri, @RequestParam(name = "schemeIris", required = false) List<String> schemeIris, @RequestParam(name = "page", required = false) Integer page, @RequestParam(name = "size", required = false) Integer size, @RequestParam(name = "inactive", required = false) boolean inactive) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.SuperiorPropertiesPaged.GET")) {
       LOG.debug("getSuperiorPropertiesPaged");
       if (null == page) page = 1;
       if (null == size) size = EntityService.MAX_CHILDREN;
       if (null == schemeIris) schemeIris = new ArrayList<>(Arrays.asList(IM.NAMESPACE, SNOMED.NAMESPACE));
-      return entityService.getSuperiorPropertiesPaged(iri, schemeIris, page, size, inactive);
+      return conceptService.getSuperiorPropertiesPaged(iri, schemeIris, page, size, inactive);
     }
   }
 
   @GetMapping(value = "/public/superiorPropertiesBoolFocusPaged")
-  @Operation(
-    summary = "Get top level properties for an entity as a tree node",
-    description = "Finds the highest parent (superior) properties for an entity and returns then in a tree node format for use in a hierarchy tree"
-  )
-  public Pageable<EntityReferenceNode> getSuperiorPropertiesBoolFocusPaged(
-    @RequestParam(name = "conceptIris") List<String> conceptIris,
-    @RequestParam(name = "schemeIris", required = false) List<String> schemeIris,
-    @RequestParam(name = "page", required = false) Integer page,
-    @RequestParam(name = "size", required = false) Integer size,
-    @RequestParam(name = "inactive", required = false) boolean inactive
-  ) throws IOException {
+  @Operation(summary = "Get top level properties for an entity as a tree node", description = "Finds the highest parent (superior) properties for an entity and returns then in a tree node format for use in a hierarchy tree")
+  public Pageable<EntityReferenceNode> getSuperiorPropertiesBoolFocusPaged(@RequestParam(name = "conceptIris") List<String> conceptIris, @RequestParam(name = "schemeIris", required = false) List<String> schemeIris, @RequestParam(name = "page", required = false) Integer page, @RequestParam(name = "size", required = false) Integer size, @RequestParam(name = "inactive", required = false) boolean inactive) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.SuperiorPropertiesBoolFocusPaged.GET")) {
       LOG.debug("getSuperiorPropertiesBoolFocusPaged");
       if (null == page) page = 1;
       if (null == size) size = EntityService.MAX_CHILDREN;
       if (null == schemeIris) schemeIris = new ArrayList<>(Arrays.asList(IM.NAMESPACE, SNOMED.NAMESPACE));
-      return entityService.getSuperiorPropertiesBoolFocusPaged(conceptIris, schemeIris, page, size, inactive);
+      return conceptService.getSuperiorPropertiesBoolFocusPaged(conceptIris, schemeIris, page, size, inactive);
     }
   }
 
   @GetMapping(value = "/public/superiorPropertyValuesPaged")
-  @Operation(
-    summary = "Get top level property values for an entity as a tree node",
-    description = "Finds the highest parent (superior) property value for an entity and returns then in a tree node format for use in a hierarchy tree"
-  )
-  public Pageable<EntityReferenceNode> getSuperiorPropertyValuesPaged(
-    @RequestParam(name = "propertyIri") String iri,
-    @RequestParam(name = "schemeIris", required = false) List<String> schemeIris,
-    @RequestParam(name = "page", required = false) Integer page,
-    @RequestParam(name = "size", required = false) Integer size,
-    @RequestParam(name = "inactive", required = false) boolean inactive
-  ) throws IOException {
+  @Operation(summary = "Get top level property values for an entity as a tree node", description = "Finds the highest parent (superior) property value for an entity and returns then in a tree node format for use in a hierarchy tree")
+  public Pageable<EntityReferenceNode> getSuperiorPropertyValuesPaged(@RequestParam(name = "propertyIri") String iri, @RequestParam(name = "schemeIris", required = false) List<String> schemeIris, @RequestParam(name = "page", required = false) Integer page, @RequestParam(name = "size", required = false) Integer size, @RequestParam(name = "inactive", required = false) boolean inactive) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.SuperiorPropertyValuesPaged.GET")) {
       LOG.debug("getSuperiorPropertyValuesPaged");
       if (null == page) page = 1;
       if (null == size) size = EntityService.MAX_CHILDREN;
       if (null == schemeIris) schemeIris = new ArrayList<>(Arrays.asList(IM.NAMESPACE, SNOMED.NAMESPACE));
-      return entityService.getSuperiorPropertyValuesPaged(iri, schemeIris, page, size, inactive);
+      return conceptService.getSuperiorPropertyValuesPaged(iri, schemeIris, page, size, inactive);
     }
   }
 
@@ -124,7 +95,7 @@ public class ConceptController {
   public List<ConceptContextMap> getConceptContextMaps(@RequestParam(name = "iri") String iri) throws IOException {
     try (MetricsTimer t = MetricsHelper.recordTime("API.Entity.ConceptContextMaps.GET")) {
       LOG.debug("getConceptContextMaps");
-      return entityService.getConceptContextMaps(iri);
+      return conceptService.getConceptContextMaps(iri);
     }
   }
 }
