+++ api/src/main/java/org/endeavourhealth/imapi/logic/ServiceHelper.java
@@ -1,0 +1,15 @@
+package org.endeavourhealth.imapi.logic;
+
+import org.endeavourhealth.imapi.model.tripletree.TTArray;
+import org.endeavourhealth.imapi.model.tripletree.TTBundle;
+import org.endeavourhealth.imapi.model.tripletree.TTValue;
+import org.endeavourhealth.imapi.vocabulary.IM;
+
+import java.util.List;
+
+import static org.endeavourhealth.imapi.model.tripletree.TTIriRef.iri;
+
+public class ServiceHelper {
+
+
+}
