+++ api/src/main/java/org/endeavourhealth/imapi/logic/exporters/ExcelSetExporter.java
@@ -5,8 +5,7 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import org.apache.poi.ss.usermodel.*;
 import org.apache.poi.xssf.usermodel.XSSFFont;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import org.endeavourhealth.imapi.dataaccess.EntityRepository2;
-import org.endeavourhealth.imapi.dataaccess.EntityTripleRepository;
+import org.endeavourhealth.imapi.dataaccess.EntityRepository;
 import org.endeavourhealth.imapi.model.exporters.SetExporterOptions;
 import org.endeavourhealth.imapi.model.iml.Concept;
 import org.endeavourhealth.imapi.model.imq.Query;
@@ -45,11 +44,10 @@ public class ExcelSetExporter {
   public static final String LEGACY_SCHEME = "Legacy scheme";
   public static final String LEGACY_USAGE = "Legacy Usage";
   public static final String CODE_ID = "code id";
-  private EntityTripleRepository entityTripleRepository = new EntityTripleRepository();
   private SetExporter setExporter = new SetExporter();
   private XSSFWorkbook workbook;
   private CellStyle headerStyle;
-  private EntityRepository2 entityRepository2 = new EntityRepository2();
+  private EntityRepository entityRepository = new EntityRepository();
 
   public ExcelSetExporter() {
     workbook = new XSSFWorkbook();
@@ -118,11 +116,11 @@ public class ExcelSetExporter {
    * @return work book
    */
   public XSSFWorkbook getSetAsExcel(SetExporterOptions options) throws JsonProcessingException, QueryException {
-    TTEntity entity = entityTripleRepository.getEntityPredicates(options.getSetIri(), Set.of(IM.DEFINITION)).getEntity();
+    TTEntity entity = entityRepository.getEntityPredicates(options.getSetIri(), Set.of(IM.DEFINITION)).getEntity();
     if (entity.getIri() == null || entity.getIri().isEmpty())
       return workbook;
 
-    String setName = entityRepository2.getBundle(options.getSetIri(), Set.of(RDFS.LABEL)).getEntity().getName();
+    String setName = entityRepository.getBundle(options.getSetIri(), Set.of(RDFS.LABEL)).getEntity().getName();
 
     String ecl = getEcl(entity);
     if (ecl != null && options.includeDefinition()) {
