+++ api/src/main/java/org/endeavourhealth/imapi/logic/reasoner/RangeInheritor.java
@@ -1,7 +1,7 @@
 package org.endeavourhealth.imapi.logic.reasoner;
 
 import org.eclipse.rdf4j.repository.RepositoryConnection;
-import org.endeavourhealth.imapi.dataaccess.EntityRepository2;
+import org.endeavourhealth.imapi.dataaccess.EntityRepository;
 
 public class RangeInheritor {
 
@@ -9,6 +9,6 @@ public class RangeInheritor {
    * Updates ranges for properties based on their super properties ranges
    */
   public void inheritRanges(RepositoryConnection conn) {
-    new EntityRepository2().inheritRanges(conn);
+    new EntityRepository().inheritRanges(conn);
   }
 }
