+++ api/src/main/java/org/endeavourhealth/imapi/logic/reasoner/SetExpander.java
@@ -1,7 +1,7 @@
 package org.endeavourhealth.imapi.logic.reasoner;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
-import org.endeavourhealth.imapi.dataaccess.EntityTripleRepository;
+import org.endeavourhealth.imapi.dataaccess.EntityRepository;
 import org.endeavourhealth.imapi.dataaccess.SetRepository;
 import org.endeavourhealth.imapi.model.iml.Concept;
 import org.endeavourhealth.imapi.model.imq.Query;
@@ -18,7 +18,7 @@ import static org.endeavourhealth.imapi.model.tripletree.TTIriRef.iri;
 
 public class SetExpander {
   private static final Logger LOG = LoggerFactory.getLogger(SetExpander.class);
-  private final EntityTripleRepository entityTripleRepository = new EntityTripleRepository();
+  private final EntityRepository entityRepository = new EntityRepository();
   private final SetRepository setRepo = new SetRepository();
 
   public void expandAllSets() throws JsonProcessingException, QueryException {
@@ -29,7 +29,7 @@ public class SetExpander {
     for (String iri : sets) {
       LOG.info("Updating members of {}", iri);
       //get the definition
-      TTBundle setDefinition = entityTripleRepository.getEntityPredicates(iri, Set.of(IM.DEFINITION));
+      TTBundle setDefinition = entityRepository.getEntityPredicates(iri, Set.of(IM.DEFINITION));
       //get the expansion.
       Set<Concept> members = setRepo.getSetExpansion(setDefinition.getEntity().get(iri(IM.DEFINITION)).asLiteral().objectValue(Query.class), false, null, List.of());
       setRepo.updateMembers(iri, members);
@@ -41,7 +41,7 @@ public class SetExpander {
 
   public void expandSet(String iri) throws QueryException, JsonProcessingException {
     LOG.info("Updating members of {}", iri);
-    TTBundle setDefinition = entityTripleRepository.getEntityPredicates(iri, Set.of(IM.DEFINITION));
+    TTBundle setDefinition = entityRepository.getEntityPredicates(iri, Set.of(IM.DEFINITION));
     if (setDefinition.getEntity().get(iri(IM.DEFINITION)) == null)
       return;
     //get the expansion.
