+++ api/src/main/java/org/endeavourhealth/imapi/logic/service/EntityService.java
@@ -7,6 +7,7 @@ import org.endeavourhealth.imapi.config.ConfigManager;
 import org.endeavourhealth.imapi.dataaccess.*;
 import org.endeavourhealth.imapi.dataaccess.helpers.XlsHelper;
 import org.endeavourhealth.imapi.filer.TTFilerException;
+import org.endeavourhealth.imapi.logic.CachedObjectMapper;
 import org.endeavourhealth.imapi.logic.exporters.ExcelSetExporter;
 import org.endeavourhealth.imapi.logic.exporters.SetExporter;
 import org.endeavourhealth.imapi.model.*;
@@ -26,7 +27,6 @@ import org.endeavourhealth.imapi.model.search.SearchResponse;
 import org.endeavourhealth.imapi.model.search.SearchResultSummary;
 import org.endeavourhealth.imapi.model.search.SearchTermCode;
 import org.endeavourhealth.imapi.model.tripletree.*;
-import org.endeavourhealth.imapi.validators.EntityValidator;
 import org.endeavourhealth.imapi.vocabulary.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -44,19 +44,14 @@ public class EntityService {
   public static final int MAX_CHILDREN = 200;
   private static final Logger LOG = LoggerFactory.getLogger(EntityService.class);
   private static final FilerService filerService = new FilerService();
-  private EntityRepository entityRepository = new EntityRepository();
-  private EntityTctRepository entityTctRepository = new EntityTctRepository();
-  private EntityTripleRepository entityTripleRepository = new EntityTripleRepository();
-  private EntityTypeRepository entityTypeRepository = new EntityTypeRepository();
+  private static DataModelService dataModelService = new DataModelService();
+  private static ConceptService conceptService = new ConceptService();
+  private static EntityRepository entityRepository = new EntityRepository();
   private ConfigManager configManager = new ConfigManager();
-  private EntityRepository2 entityRepository2 = new EntityRepository2();
-  private SearchService searchService = new SearchService();
 
   private static void filterOutSpecifiedPredicates(Set<String> excludePredicates, TTBundle bundle) {
     if (excludePredicates != null) {
-      Map<String, String> filtered = bundle.getPredicates().entrySet().stream()
-        .filter(entry -> !entry.getKey().equals(RDFS.LABEL) && entry.getValue() != null)
-        .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
+      Map<String, String> filtered = bundle.getPredicates().entrySet().stream().filter(entry -> !entry.getKey().equals(RDFS.LABEL) && entry.getValue() != null).collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
       bundle.setPredicates(filtered);
       if (excludePredicates.contains(RDFS.LABEL)) {
         bundle.getEntity().set(iri(RDFS.LABEL), (TTValue) null);
@@ -64,48 +59,28 @@ public class EntityService {
     }
   }
 
-  private static void filterOutInactiveTermCodes(TTBundle bundle) {
-    if (bundle.getEntity().get(iri(IM.HAS_TERM_CODE)) != null) {
-      List<TTValue> termCodes = bundle.getEntity().get(iri(IM.HAS_TERM_CODE)).getElements();
-      TTArray activeTermCodes = new TTArray();
-      for (TTValue value : termCodes) {
-        if (value.asNode().get(iri(IM.HAS_STATUS)) != null) {
-          if (value.asNode().get(iri(IM.HAS_STATUS)) != null && "Active".equals(value.asNode().get(iri(IM.HAS_STATUS)).asIriRef().getName())) {
-            activeTermCodes.add(value);
-          }
-        } else
-          activeTermCodes.add(value);
-      }
-      bundle.getEntity().set(iri(IM.HAS_TERM_CODE), activeTermCodes);
-    }
-  }
-
-  public TTBundle getBundle(String iri, Set<String> predicates) {
-    return entityRepository2.getBundle(iri, predicates);
+  public static TTBundle getBundle(String iri, Set<String> predicates) {
+    return entityRepository.getBundle(iri, predicates);
   }
 
   public TTBundle getBundleByPredicateExclusions(String iri, Set<String> excludePredicates) {
-    TTBundle bundle = entityRepository2.getBundle(iri, excludePredicates, true);
+    TTBundle bundle = entityRepository.getBundle(iri, excludePredicates, true);
     filterOutSpecifiedPredicates(excludePredicates, bundle);
     filterOutInactiveTermCodes(bundle);
     return bundle;
   }
 
   public TTIriRef getEntityReference(String iri) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
     return entityRepository.getEntityReferenceByIri(iri);
   }
 
-  public List<EntityReferenceNode> getImmediateChildren(String iri, List<String> schemeIris, Integer pageIndex, Integer pageSize,
-                                                        boolean inactive) {
-    if (iri == null || iri.isEmpty())
-      return Collections.emptyList();
+  public static List<EntityReferenceNode> getImmediateChildren(String iri, List<String> schemeIris, Integer pageIndex, Integer pageSize, boolean inactive) {
+    if (iri == null || iri.isEmpty()) return Collections.emptyList();
 
     List<EntityReferenceNode> result = new ArrayList<>();
     int rowNumber = 0;
-    if (pageIndex != null && pageSize != null)
-      rowNumber = (pageIndex - 1) * pageSize;
+    if (pageIndex != null && pageSize != null) rowNumber = (pageIndex - 1) * pageSize;
 
     for (TTIriRef c : getChildren(iri, schemeIris, rowNumber, pageSize, inactive)) {
       result.add(getEntityAsEntityReferenceNode(c.getIri()));
@@ -116,92 +91,76 @@ public class EntityService {
     return result;
   }
 
-  public EntityReferenceNode getEntityAsEntityReferenceNode(String iri) {
+  public static EntityReferenceNode getEntityAsEntityReferenceNode(String iri) {
     return getEntityAsEntityReferenceNode(iri, null, false);
   }
 
-  public EntityReferenceNode getEntityAsEntityReferenceNode(String iri, List<String> schemeIris, boolean inactive) {
+  public static EntityReferenceNode getEntityAsEntityReferenceNode(String iri, List<String> schemeIris, boolean inactive) {
     if (null == iri) throw new IllegalArgumentException("Missing iri parameter");
 
-    return entityTripleRepository.getEntityReferenceNode(iri, schemeIris, inactive);
+    return entityRepository.getEntityReferenceNode(iri, schemeIris, inactive);
   }
 
   public Pageable<EntityReferenceNode> getEntityChildrenPagedWithTotalCount(String iri, List<String> schemeIris, Integer page, Integer size, boolean inactive) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
 
     int rowNumber = 0;
-    if (page != null && size != null)
-      rowNumber = (page - 1) * size;
+    if (page != null && size != null) rowNumber = (page - 1) * size;
 
-    Pageable<TTIriRef> childrenAndTotalCount = entityTripleRepository.findImmediateChildrenPagedByIriWithTotalCount(iri, schemeIris, rowNumber, size, inactive);
+    Pageable<TTIriRef> childrenAndTotalCount = entityRepository.findImmediateChildrenPagedByIriWithTotalCount(iri, schemeIris, rowNumber, size, inactive);
     return iriRefPageableToEntityReferenceNodePageable(childrenAndTotalCount, schemeIris, inactive);
   }
 
   public Pageable<TTIriRef> getPartialWithTotalCount(String iri, String predicateList, List<String> schemeIris, Integer page, Integer size, boolean inactive) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
 
     int rowNumber = 0;
-    if (page != null && size != null)
-      rowNumber = (page - 1) * size;
+    if (page != null && size != null) rowNumber = (page - 1) * size;
 
-    return entityTripleRepository.findPartialWithTotalCount(iri, predicateList, schemeIris, rowNumber, size, inactive);
+    return entityRepository.findPartialWithTotalCount(iri, predicateList, schemeIris, rowNumber, size, inactive);
   }
 
-  private List<TTIriRef> getChildren(String iri, List<String> schemeIris, int rowNumber, Integer pageSize, boolean inactive) {
-    return entityTripleRepository.findImmediateChildrenByIri(iri, schemeIris, rowNumber, pageSize, inactive);
+  private static List<TTIriRef> getChildren(String iri, List<String> schemeIris, int rowNumber, Integer pageSize, boolean inactive) {
+    return entityRepository.findImmediateChildrenByIri(iri, schemeIris, rowNumber, pageSize, inactive);
   }
 
-  public List<EntityReferenceNode> getImmediateParents(String iri, List<String> schemeIris, Integer pageIndex, Integer pageSize,
-                                                       boolean inactive) {
+  public List<EntityReferenceNode> getImmediateParents(String iri, List<String> schemeIris, Integer pageIndex, Integer pageSize, boolean inactive) {
 
-    if (iri == null || iri.isEmpty())
-      return Collections.emptyList();
+    if (iri == null || iri.isEmpty()) return Collections.emptyList();
 
     int rowNumber = 0;
-    if (pageIndex != null && pageSize != null)
-      rowNumber = (pageIndex - 1) * pageSize;
+    if (pageIndex != null && pageSize != null) rowNumber = (pageIndex - 1) * pageSize;
 
-    List<EntityReferenceNode> parents = getParents(iri, schemeIris, rowNumber, pageSize, inactive).stream()
-      .map(p -> new EntityReferenceNode(p.getIri(), p.getName())).toList();
+    List<EntityReferenceNode> parents = getParents(iri, schemeIris, rowNumber, pageSize, inactive).stream().map(p -> new EntityReferenceNode(p.getIri(), p.getName())).toList();
 
     for (EntityReferenceNode parent : parents)
-      parent.setType(entityTypeRepository.getEntityTypes(parent.getIri()));
+      parent.setType(entityRepository.getEntityTypes(parent.getIri()));
 
     return parents;
   }
 
   private List<TTIriRef> getParents(String iri, List<String> schemeIris, int rowNumber, Integer pageSize, boolean inactive) {
 
-    return entityTripleRepository.findImmediateParentsByIri(iri, schemeIris, rowNumber, pageSize, inactive);
+    return entityRepository.findImmediateParentsByIri(iri, schemeIris, rowNumber, pageSize, inactive);
   }
 
   public List<TTIriRef> isWhichType(String iri, List<String> candidates) {
-    if (iri == null || iri.isEmpty() || candidates == null || candidates.isEmpty())
-      return Collections.emptyList();
-    return entityTctRepository
-      .findAncestorsByType(iri, RDFS.SUBCLASS_OF, candidates).stream()
-      .sorted(Comparator.comparing(TTIriRef::getName)).toList();
+    if (iri == null || iri.isEmpty() || candidates == null || candidates.isEmpty()) return Collections.emptyList();
+    return entityRepository.findAncestorsByType(iri, RDFS.SUBCLASS_OF, candidates).stream().sorted(Comparator.comparing(TTIriRef::getName)).toList();
   }
 
   public List<TTEntity> usages(String iri, Integer pageIndex, Integer pageSize) throws JsonProcessingException {
     ArrayList<TTEntity> usageEntities = new ArrayList<>();
-    if (iri == null || iri.isEmpty())
-      return Collections.emptyList();
+    if (iri == null || iri.isEmpty()) return Collections.emptyList();
 
     List<String> xmlDataTypes = configManager.getConfig(CONFIG.XML_SCHEMA_DATA_TYPES, new TypeReference<>() {
     });
-    if (xmlDataTypes != null && xmlDataTypes.contains(iri))
-      return Collections.emptyList();
+    if (xmlDataTypes != null && xmlDataTypes.contains(iri)) return Collections.emptyList();
 
     int rowNumber = 0;
-    if (pageIndex != null && pageSize != null)
-      rowNumber = pageIndex * pageSize;
+    if (pageIndex != null && pageSize != null) rowNumber = pageIndex * pageSize;
 
-    List<TTIriRef> usageRefs = entityTripleRepository.getConceptUsages(iri, rowNumber, pageSize).stream()
-      .sorted(Comparator.comparing(TTIriRef::getName, Comparator.nullsLast(Comparator.naturalOrder())))
-      .distinct().toList();
+    List<TTIriRef> usageRefs = entityRepository.getConceptUsages(iri, rowNumber, pageSize).stream().sorted(Comparator.comparing(TTIriRef::getName, Comparator.nullsLast(Comparator.naturalOrder()))).distinct().toList();
 
     usageRefs = usageRefs.stream().filter(usage -> !usage.getIri().equals(iri)).toList();
     for (TTIriRef usage : usageRefs) {
@@ -213,50 +172,13 @@ public class EntityService {
   }
 
   public Integer totalRecords(String iri) throws JsonProcessingException {
-    if (iri == null || iri.isEmpty())
-      return 0;
+    if (iri == null || iri.isEmpty()) return 0;
 
     List<String> xmlDataTypes = configManager.getConfig(CONFIG.XML_SCHEMA_DATA_TYPES, new TypeReference<>() {
     });
-    if (xmlDataTypes != null && xmlDataTypes.contains(iri))
-      return 0;
+    if (xmlDataTypes != null && xmlDataTypes.contains(iri)) return 0;
 
-    return entityTripleRepository.getConceptUsagesCount(iri);
-  }
-
-  public SearchResponse advancedSearch(QueryRequest request) throws OpenSearchException, QueryException {
-    SearchResponse searchResults = searchService.getEntitiesByTerm(request);
-    searchResults.setPage(request.getPage().getPageNumber());
-    return searchResults;
-  }
-
-  public List<SearchTermCode> getEntityTermCodes(String iri, boolean includeInactive) {
-    if (iri == null || iri.isEmpty())
-      return Collections.emptyList();
-    TTBundle termsBundle = getBundle(iri, Stream.of(IM.HAS_TERM_CODE).collect(Collectors.toSet()));
-    if (!includeInactive) filterOutInactiveTermCodes(termsBundle);
-    TTArray terms = termsBundle.getEntity().get(iri(IM.HAS_TERM_CODE));
-    if (null == terms) return Collections.emptyList();
-    List<SearchTermCode> termsSummary = new ArrayList<>();
-    for (TTValue term : terms.getElements()) {
-      processTerm(term, termsSummary);
-    }
-    return termsSummary;
-  }
-
-  private void processTerm(TTValue term, List<SearchTermCode> termsSummary) {
-    if (null != term.asNode().get(iri(IM.CODE)) && null == termsSummary.stream().filter(t -> term.asNode().get(iri(IM.CODE)).get(0).asLiteral().getValue().equals(t.getCode())).findAny().orElse(null)) {
-      SearchTermCode newTerm = new SearchTermCode();
-      if (term.asNode().has(iri(IM.CODE)))
-        newTerm.setCode(term.asNode().get(iri(IM.CODE)).get(0).asLiteral().getValue());
-      if (term.asNode().has(iri(RDFS.LABEL)))
-        newTerm.setTerm(term.asNode().get(iri(RDFS.LABEL)).get(0).asLiteral().getValue());
-      if (term.asNode().has(iri(IM.HAS_STATUS)))
-        newTerm.setStatus(term.asNode().get(iri(IM.HAS_STATUS)).get(0).asIriRef());
-      termsSummary.add(
-        newTerm
-      );
-    }
+    return entityRepository.getConceptUsagesCount(iri);
   }
 
   public TTEntity getSummaryFromConfig(String iri, List<ComponentLayoutItem> configs) {
@@ -270,8 +192,7 @@ public class EntityService {
   }
 
   public DownloadDto getJsonDownload(String iri, List<ComponentLayoutItem> configs, DownloadEntityOptions params) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
 
     DownloadDto downloadDto = new DownloadDto();
 
@@ -284,10 +205,10 @@ public class EntityService {
       downloadDto.setInferred(getBundle(iri, new HashSet<>(Arrays.asList(RDFS.SUBCLASS_OF, IM.ROLE_GROUP))).getEntity());
     }
     if (params.includeProperties()) {
-      downloadDto.setDataModelProperties(getDataModelProperties(iri));
+      downloadDto.setDataModelProperties(dataModelService.getDataModelProperties(iri));
     }
     if (params.includeTerms()) {
-      downloadDto.setTerms(getEntityTermCodes(iri, false));
+      downloadDto.setTerms(conceptService.getEntityTermCodes(iri, false));
     }
     if (params.includeIsChildOf()) {
       downloadDto.setIsChildOf(getBundle(iri, new HashSet<>(List.of(IM.IS_CHILD_OF))).getEntity().get(iri(IM.IS_CHILD_OF)));
@@ -300,8 +221,7 @@ public class EntityService {
   }
 
   public XlsHelper getExcelDownload(String iri, List<ComponentLayoutItem> configs, DownloadEntityOptions params) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
 
     XlsHelper xls = new XlsHelper();
 
@@ -314,10 +234,10 @@ public class EntityService {
       xls.addInferred(getBundle(iri, new HashSet<>(Arrays.asList(RDFS.SUBCLASS_OF, IM.ROLE_GROUP))));
     }
     if (params.includeProperties()) {
-      xls.addDataModelProperties(getDataModelProperties(iri));
+      xls.addDataModelProperties(dataModelService.getDataModelProperties(iri));
     }
     if (params.includeTerms()) {
-      xls.addTerms(getEntityTermCodes(iri, false));
+      xls.addTerms(conceptService.getEntityTermCodes(iri, false));
     }
     TTEntity isChildOfEntity = getBundle(iri, new HashSet<>(List.of(IM.IS_CHILD_OF))).getEntity();
     TTArray isChildOfData = isChildOfEntity.get(iri(IM.IS_CHILD_OF));
@@ -333,99 +253,6 @@ public class EntityService {
     return xls;
   }
 
-  public TTEntity getDataModelPropertiesAndSubClasses(String iri, String parent) {
-    TTEntity entity = entityTripleRepository.getEntityPredicates(iri, Set.of(SHACL.PROPERTY)).getEntity();
-    List<TTValue> properties = new ArrayList<>();
-    if (entity.get(iri(SHACL.PROPERTY)) != null) {
-      for (TTValue property : entity.get(iri(SHACL.PROPERTY)).getElements()) {
-        if (property.asNode().get(iri(IM.INHERITED_FROM)) == null || parent == null)
-          properties.add(property);
-        else if (!property.asNode().get(iri(IM.INHERITED_FROM)).asIriRef().getIri().equals(parent))
-          properties.add(property);
-      }
-    }
-    ;
-    List<EntityReferenceNode> children = getImmediateChildren(iri, null, null, null, false);
-    if (children != null && !children.isEmpty()) {
-      TTNode subclasses = new TTNode()
-        .set(iri(SHACL.ORDER), TTLiteral.literal(0))
-        .set(iri(SHACL.PATH), new TTIriRef().setIri(IM.NAMESPACE + "hasSubclasses").setName("has subtypes"));
-      for (EntityReferenceNode type : children) {
-        subclasses.addObject(iri(SHACL.NODE), new TTIriRef().setIri(type.getIri()).setName(type.getName()));
-      }
-      properties.add(subclasses);
-    }
-    if (!properties.isEmpty()) {
-      properties = properties
-        .stream().sorted(Comparator
-          .comparing(p -> p.asNode().get(iri(SHACL.ORDER)) == null ? 1 : p.asNode().get(iri(SHACL.ORDER)).asLiteral().intValue())).toList();
-      TTArray ordered = new TTArray();
-      for (TTValue property : properties)
-        ordered.add(property);
-      entity.set(iri(SHACL.PROPERTY), ordered);
-      return entity;
-    } else {
-      return entity;
-    }
-  }
-
-  public List<DataModelProperty> getDataModelProperties(String iri) {
-    TTEntity entity = getBundle(iri, Set.of(SHACL.PROPERTY, RDFS.LABEL)).getEntity();
-    return getDataModelProperties(entity);
-  }
-
-  public List<DataModelProperty> getDataModelProperties(TTEntity entity) {
-    List<DataModelProperty> properties = new ArrayList<>();
-    if (entity == null)
-      return Collections.emptyList();
-    if (entity.has(iri(SHACL.PROPERTY))) {
-      getDataModelPropertyGroups(entity, properties);
-    }
-    return properties.stream().sorted(Comparator.comparing(DataModelProperty::getOrder)).toList();
-  }
-
-  private void getDataModelPropertyGroups(TTEntity entity, List<DataModelProperty> properties) {
-    for (TTValue propertyGroup : entity.get(iri(SHACL.PROPERTY)).iterator()) {
-      if (propertyGroup.isNode()) {
-        TTIriRef inheritedFrom = propertyGroup.asNode().has(iri(IM.INHERITED_FROM))
-          ? propertyGroup.asNode().get(iri(IM.INHERITED_FROM)).asIriRef()
-          : null;
-        if (propertyGroup.asNode().has(iri(SHACL.PATH))) {
-          getDataModelShaclProperties(properties, propertyGroup, inheritedFrom);
-        }
-      }
-    }
-  }
-
-  private void getDataModelShaclProperties(List<DataModelProperty> properties, TTValue propertyGroup, TTIriRef inheritedFrom) {
-    TTIriRef propertyPath = propertyGroup.asNode().get(iri(SHACL.PATH)).asIriRef();
-    if (properties.stream()
-      .noneMatch(o -> o.getProperty().getIri().equals(propertyPath.getIri()))) {
-      properties.add(getPropertyValue(inheritedFrom, propertyGroup, propertyPath));
-    }
-  }
-
-  private DataModelProperty getPropertyValue(TTIriRef inheritedFrom, TTValue property, TTIriRef propertyPath) {
-    DataModelProperty pv = new DataModelProperty().setInheritedFrom(inheritedFrom).setProperty(propertyPath);
-
-    if (property.asNode().has(iri(SHACL.CLASS)))
-      pv.setType(property.asNode().get(iri(SHACL.CLASS)).asIriRef());
-    if (property.asNode().has(iri(SHACL.NODE)))
-      pv.setType(property.asNode().get(iri(SHACL.NODE)).asIriRef());
-    if (property.asNode().has(iri(OWL.CLASS)))
-      pv.setType(property.asNode().get(iri(OWL.CLASS)).asIriRef());
-    if (property.asNode().has(iri(SHACL.DATATYPE)))
-      pv.setType(property.asNode().get(iri(SHACL.DATATYPE)).asIriRef());
-    if (property.asNode().has(iri(SHACL.FUNCTION)))
-      pv.setType(property.asNode().get(iri(SHACL.FUNCTION)).asIriRef());
-    if (property.asNode().has(iri(SHACL.MAXCOUNT)))
-      pv.setMaxExclusive(property.asNode().get(iri(SHACL.MAXCOUNT)).asLiteral().getValue());
-    if (property.asNode().has(iri(SHACL.MINCOUNT)))
-      pv.setMinExclusive(property.asNode().get(iri(SHACL.MINCOUNT)).asLiteral().getValue());
-    pv.setOrder(property.asNode().has(iri(SHACL.ORDER)) ? property.asNode().get(iri(SHACL.ORDER)).asLiteral().intValue() : 0);
-
-    return pv;
-  }
 
   public GraphDto getGraphData(String iri) {
     TTEntity entity = getBundle(iri, Set.of(RDFS.SUBCLASS_OF, RDFS.LABEL)).getEntity();
@@ -437,19 +264,11 @@ public class EntityService {
     TTBundle axioms = getBundle(iri, new HashSet<>(Arrays.asList(RDFS.SUBCLASS_OF, IM.ROLE_GROUP)));
     List<GraphDto> axiomGraph = new ArrayList<>();
 
-    List<GraphDto> dataModelProps = getDataModelProperties(iri).stream()
-      .map(prop -> new GraphDto(prop.getProperty().getIri(), prop.getProperty().getName(),
-        prop.getType() != null ? prop.getType().getIri() : "",
-        prop.getType() != null ? prop.getType().getName() : "",
-        prop.getInheritedFrom() != null ? prop.getInheritedFrom().getIri() : "",
-        prop.getInheritedFrom() != null ? prop.getInheritedFrom().getName() : ""))
-      .toList();
+    List<GraphDto> dataModelProps = dataModelService.getDataModelProperties(iri).stream().map(prop -> new GraphDto(prop.getProperty().getIri(), prop.getProperty().getName(), prop.getType() != null ? prop.getType().getIri() : "", prop.getType() != null ? prop.getType().getName() : "", prop.getInheritedFrom() != null ? prop.getInheritedFrom().getIri() : "", prop.getInheritedFrom() != null ? prop.getInheritedFrom().getName() : "")).toList();
 
     List<GraphDto> isas = getEntityDefinedParents(entity);
 
-    List<GraphDto> subtypes = getDefinitionSubTypes(iri).stream()
-      .map(subtype -> new GraphDto().setName(subtype.getName()).setIri(subtype.getIri()))
-      .toList();
+    List<GraphDto> subtypes = getDefinitionSubTypes(iri).stream().map(subtype -> new GraphDto().setName(subtype.getName()).setIri(subtype.getIri())).toList();
 
     GraphDto axiom = new GraphDto().setKey("0_2").setName("Axioms");
     GraphDto axiomWrapper = getWrapper(axiomGraph, "0_2_0");
@@ -472,12 +291,10 @@ public class EntityService {
       dataModel.getChildren().add(dataModelInherited);
     }
 
-    GraphDto childrenWrapper = new GraphDto().setKey("0_1_0")
-      .setType(!subtypes.isEmpty() ? GraphType.SUBTYPE : GraphType.NONE);
+    GraphDto childrenWrapper = new GraphDto().setKey("0_1_0").setType(!subtypes.isEmpty() ? GraphType.SUBTYPE : GraphType.NONE);
     childrenWrapper.getLeafNodes().addAll(subtypes);
 
-    GraphDto parentsWrapper = new GraphDto().setKey("0_0_0")
-      .setType(!isas.isEmpty() ? GraphType.ISA : GraphType.NONE);
+    GraphDto parentsWrapper = new GraphDto().setKey("0_0_0").setType(!isas.isEmpty() ? GraphType.ISA : GraphType.NONE);
     parentsWrapper.getLeafNodes().addAll(isas);
 
     graphParents.getChildren().add(parentsWrapper);
@@ -496,56 +313,43 @@ public class EntityService {
 
   private GraphDto getWrapper(List<GraphDto> props, String key) {
     GraphDto wrapper = new GraphDto().setKey(key).setType(GraphType.PROPERTIES);
-    wrapper.getLeafNodes()
-      .addAll(props.stream()
-        .filter(prop -> prop.getInheritedFromIri() == null).toList());
+    wrapper.getLeafNodes().addAll(props.stream().filter(prop -> prop.getInheritedFromIri() == null).toList());
     return wrapper;
   }
 
   private GraphDto getDataModelInheritedWrapper(List<GraphDto> dataModelProps) {
     GraphDto dataModelInheritedWrapper = new GraphDto().setKey("0_3_1_0").setType(GraphType.PROPERTIES);
-    dataModelInheritedWrapper.getLeafNodes()
-      .addAll(dataModelProps.stream()
-        .filter(prop -> prop.getInheritedFromIri() != null).toList());
+    dataModelInheritedWrapper.getLeafNodes().addAll(dataModelProps.stream().filter(prop -> prop.getInheritedFromIri() != null).toList());
     return dataModelInheritedWrapper;
   }
 
   private void addWrapper(GraphDto dto, GraphDto wrapper, String key) {
-    dto.getChildren()
-      .add(wrapper.getLeafNodes().isEmpty()
-        ? new GraphDto().setKey(key).setType(GraphType.NONE)
-        : wrapper);
+    dto.getChildren().add(wrapper.getLeafNodes().isEmpty() ? new GraphDto().setKey(key).setType(GraphType.NONE) : wrapper);
   }
 
   private List<GraphDto> getEntityDefinedParents(TTEntity entity) {
     TTArray parent = entity.get(iri(RDFS.SUBCLASS_OF));
-    if (parent == null)
-      return Collections.emptyList();
+    if (parent == null) return Collections.emptyList();
     List<GraphDto> result = new ArrayList<>();
     parent.getElements().forEach(item -> {
       if (!OWL.THING.equals(item.asIriRef().getIri()))
-        result.add(new GraphDto().setIri(item.asIriRef().getIri()).setName(item.asIriRef().getName())
-          .setPropertyType(iri(RDFS.SUBCLASS_OF).getName()));
+        result.add(new GraphDto().setIri(item.asIriRef().getIri()).setName(item.asIriRef().getName()).setPropertyType(iri(RDFS.SUBCLASS_OF).getName()));
     });
 
     return result;
   }
 
   public List<TTIriRef> getDefinitionSubTypes(String iri) {
-
-    return entityTripleRepository.findImmediateChildrenByIri(iri, null, null, null, false).stream()
-      .map(t -> new TTIriRef(t.getIri(), t.getName())).toList();
+    return entityRepository.findImmediateChildrenByIri(iri, null, null, null, false).stream().map(t -> new TTIriRef(t.getIri(), t.getName())).toList();
   }
 
   public SearchResultSummary getSummary(String iri) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
     return entityRepository.getEntitySummaryByIri(iri);
   }
 
   public TTEntity getConceptShape(String iri) {
-    if (iri == null || iri.isEmpty())
-      return null;
+    if (iri == null || iri.isEmpty()) return null;
     TTEntity entity = getBundle(iri, Set.of(SHACL.PROPERTY, SHACL.OR, RDF.TYPE)).getEntity();
     TTArray value = entity.get(iri(RDF.TYPE));
     if (!value.getElements().contains(iri(SHACL.NODESHAPE))) {
@@ -555,7 +359,7 @@ public class EntityService {
   }
 
   public List<Namespace> getNamespaces() {
-    return entityTripleRepository.findNamespaces();
+    return entityRepository.findNamespaces();
   }
 
   public TTBundle getInferredBundle(String iri) {
@@ -581,7 +385,7 @@ public class EntityService {
       return null;
     }
     TTDocument document = new TTDocument();
-    List<Namespace> namespaces = entityTripleRepository.findNamespaces();
+    List<Namespace> namespaces = entityRepository.findNamespaces();
     TTContext context = new TTContext();
     for (Namespace namespace : namespaces) {
       context.add(namespace.getIri(), namespace.getPrefix(), namespace.getName());
@@ -594,42 +398,6 @@ public class EntityService {
     return document;
   }
 
-  public List<SimpleMap> getMatchedFrom(String iri) {
-    if (iri == null || iri.isEmpty()) return new ArrayList<>();
-    String scheme = iri.substring(0, iri.indexOf("#") + 1);
-    List<Namespace> namespaces = getNamespaces();
-    List<String> schemes = namespaces.stream().map(Namespace::getIri).collect(Collectors.toList());
-    schemes.remove(scheme);
-    return entityTripleRepository.getMatchedFrom(iri, schemes);
-  }
-
-  public List<SimpleMap> getMatchedTo(String iri) {
-    if (iri == null || iri.isEmpty()) return new ArrayList<>();
-    String scheme = iri.substring(0, iri.indexOf("#") + 1);
-    List<Namespace> namespaces = getNamespaces();
-    List<String> schemes = namespaces.stream().map(Namespace::getIri).collect(Collectors.toList());
-    schemes.remove(scheme);
-    return entityTripleRepository.getMatchedTo(iri, schemes);
-  }
-
-  public XSSFWorkbook getSetExport(SetExporterOptions options) throws JsonProcessingException, QueryException {
-    if (options.getSetIri() == null || options.getSetIri().isEmpty()) {
-      throw new IllegalArgumentException("SetIri is required");
-    }
-    return new ExcelSetExporter().getSetAsExcel(options);
-  }
-
-  /**
-   * Returns an entity and predicate names from an iri including all predicates and blank nodes recursively to a depth of 5
-   *
-   * @param iri the string representation of the absolute iri
-   * @return a bundle containing an entity and predicate iri to name map.
-   */
-  public TTBundle getFullEntity(String iri) {
-    return entityRepository2.getBundle(iri);
-  }
-
-
   public List<TTIriRef> getParentPath(String iri) {
     TTEntity entity = getBundle(iri, new HashSet<>(List.of(RDFS.LABEL))).getEntity();
     List<TTIriRef> parents = new ArrayList<>();
@@ -640,7 +408,7 @@ public class EntityService {
   }
 
   private void getParentPathRecursive(String iri, List<TTIriRef> parents) {
-    TTIriRef parent = entityTripleRepository.findParentFolderRef(iri);
+    TTIriRef parent = entityRepository.findParentFolderRef(iri);
     if (parent != null) {
       parents.add(parent);
       getParentPathRecursive(parent.getIri(), parents);
@@ -649,7 +417,7 @@ public class EntityService {
 
   public Set<TTIriRef> getNames(Set<String> iris) {
     Set<TTIriRef> result = iris.stream().map(TTIriRef::new).collect(Collectors.toSet());
-    entityRepository2.getNames(result);
+    entityRepository.getNames(result);
     return result;
   }
 
@@ -723,57 +491,40 @@ public class EntityService {
     return found ? i : -1;
   }
 
-  public boolean iriExists(String iri) {
+  public static boolean iriExists(String iri) {
     return entityRepository.iriExists(iri);
   }
 
-  public TTEntity createEntity(TTEntity entity, String agentName) throws TTFilerException, JsonProcessingException {
-    EntityValidator validator = new EntityValidator();
-    validator.isValid(entity, this, "Create");
+  public static TTEntity createEntity(TTEntity entity, String agentName) throws TTFilerException, JsonProcessingException {
+    isValid(entity, "Create");
     TTIriRef graph = iri(GRAPH.DISCOVERY);
     entity.setCrud(iri(IM.ADD_QUADS)).setVersion(1);
     filerService.fileEntity(entity, graph, agentName, null);
     return entity;
   }
 
-  public TTEntity updateEntity(TTEntity entity, String agentName) throws TTFilerException, JsonProcessingException {
-    EntityValidator validator = new EntityValidator();
-    validator.isValid(entity, this, "Update");
+  public static TTEntity updateEntity(TTEntity entity, String agentName) throws TTFilerException, JsonProcessingException {
+    isValid(entity, "Update");
     TTIriRef graph = iri(GRAPH.DISCOVERY);
     entity.setCrud(iri(IM.UPDATE_ALL));
-    TTEntity usedEntity = getFullEntity(entity.getIri()).getEntity();
+    TTEntity usedEntity = getBundle(entity.getIri(), null).getEntity();
     entity.setVersion(usedEntity.getVersion() + 1);
     filerService.fileEntity(entity, graph, agentName, usedEntity);
     return entity;
   }
 
-  public void updateSubsetsFromSuper(String agentName, TTEntity entity) throws TTFilerException, JsonProcessingException {
-    TTArray subsets = entity.get(iri(IM.HAS_SUBSET));
-    String entityIri = entity.getIri();
-    Set<TTIriRef> subsetsOriginal = getSubsets(entityIri);
-    List<TTIriRef> subsetsArray = subsets.stream().map(TTValue::asIriRef).toList();
-    for (TTIriRef subset : subsetsArray) {
-      TTEntity subsetEntity = getFullEntity(subset.getIri()).getEntity();
-      if (null != subsetEntity) {
-        if (!(subsetEntity.isType(iri(IM.VALUESET)) || subsetEntity.isType(iri(IM.CONCEPT_SET))))
-          throw new TTFilerException("Subsets must be of type valueSet or conceptSet. Type: " + subsetEntity.getType());
-        TTArray isSubsetOf = subsetEntity.get(iri(IM.IS_SUBSET_OF));
-        if (null == isSubsetOf) {
-          subsetEntity.set(iri(IM.IS_SUBSET_OF), new TTArray().add(iri(entityIri)));
-          updateEntity(subsetEntity, agentName);
-        } else if (isSubsetOf.getElements().stream().noneMatch(i -> Objects.equals(i.asIriRef().getIri(), entityIri))) {
-          isSubsetOf.add(iri(entityIri));
-          updateEntity(subsetEntity, agentName);
-        }
+  protected static void filterOutInactiveTermCodes(TTBundle bundle) {
+    if (bundle.getEntity().get(iri(IM.HAS_TERM_CODE)) != null) {
+      List<TTValue> termCodes = bundle.getEntity().get(iri(IM.HAS_TERM_CODE)).getElements();
+      TTArray activeTermCodes = new TTArray();
+      for (TTValue value : termCodes) {
+        if (value.asNode().get(iri(IM.HAS_STATUS)) != null) {
+          if (value.asNode().get(iri(IM.HAS_STATUS)) != null && "Active".equals(value.asNode().get(iri(IM.HAS_STATUS)).asIriRef().getName())) {
+            activeTermCodes.add(value);
+          }
+        } else activeTermCodes.add(value);
       }
-    }
-    for (TTIriRef subsetOriginal : subsetsOriginal) {
-      if (subsetsArray.stream().noneMatch(s -> s.getIri().equals(subsetOriginal.getIri()))) {
-        TTEntity subsetEntity = getFullEntity(subsetOriginal.getIri()).getEntity();
-        TTArray isSubsetOf = subsetEntity.get(iri(IM.IS_SUBSET_OF));
-        isSubsetOf.remove(iri(entityIri));
-        updateEntity(subsetEntity, agentName);
-      }
+      bundle.getEntity().set(iri(IM.HAS_TERM_CODE), activeTermCodes);
     }
   }
 
@@ -789,61 +540,18 @@ public class EntityService {
     return entityRepository.getOSDocument(iri);
   }
 
-  public List<TTIriRef> getProperties() {
-    return entityRepository.getProperties();
-  }
-
-  public List<TTIriRef> getDistillation(List<TTIriRef> conceptList) {
-    List<String> iriList = conceptList.stream().map(c -> "<" + c.getIri() + ">").toList();
-    String iris = String.join(" ", iriList);
-    Set<String> isas = entityRepository.getDistillation(iris);
-    conceptList.removeIf(c -> isas.contains(c.getIri()));
-    return conceptList;
-  }
-
   public Set<String> getPredicates(String iri) {
     return entityRepository.getPredicates(iri);
   }
 
-  public Pageable<EntityReferenceNode> getSuperiorPropertiesPaged(String iri, List<String> schemeIris, Integer page, Integer size, boolean inactive) {
-    if (null == iri || iri.isEmpty()) return null;
-
-    int rowNumber = 0;
-    if (null != page && null != size) rowNumber = (page - 1) * size;
-
-    Pageable<TTIriRef> propertiesAndCount = entityTripleRepository.getSuperiorPropertiesByConceptPagedWithTotalCount(iri, rowNumber, size, inactive);
-    return iriRefPageableToEntityReferenceNodePageable(propertiesAndCount, schemeIris, inactive);
-  }
-
-  public Pageable<EntityReferenceNode> getSuperiorPropertiesBoolFocusPaged(List<String> conceptIris, List<String> schemeIris, Integer page, Integer size, boolean inactive) {
-    if (null == conceptIris || conceptIris.isEmpty()) return null;
-
-    int rowNumber = 0;
-    if (null != page && null != size) rowNumber = (page - 1) * size;
-
-    Pageable<TTIriRef> propertiesAndCount = entityTripleRepository.getSuperiorPropertiesByConceptBoolFocusPagedWithTotalCount(conceptIris, rowNumber, size, inactive);
-    return iriRefPageableToEntityReferenceNodePageable(propertiesAndCount, schemeIris, inactive);
-  }
-
-  public Pageable<EntityReferenceNode> getSuperiorPropertyValuesPaged(String iri, List<String> schemeIris, Integer page, Integer size, boolean inactive) {
-    if (null == iri || iri.isEmpty()) return null;
-
-    int rowNumber = 0;
-    if (null != page && null != size) rowNumber = (page - 1) * size;
-
-
-    Pageable<TTIriRef> propertiesAndCount = entityTripleRepository.getSuperiorPropertyValuesByPropertyPagedWithTotalCount(iri, rowNumber, size, inactive);
-    return iriRefPageableToEntityReferenceNodePageable(propertiesAndCount, schemeIris, inactive);
-  }
-
-  private Pageable<EntityReferenceNode> iriRefPageableToEntityReferenceNodePageable(Pageable<TTIriRef> iriRefPageable, List<String> schemeIris, boolean inactive) {
+  protected static Pageable<EntityReferenceNode> iriRefPageableToEntityReferenceNodePageable(Pageable<TTIriRef> iriRefPageable, List<String> schemeIris, boolean inactive) {
     Pageable<EntityReferenceNode> result = new Pageable<>();
     result.setTotalCount(iriRefPageable.getTotalCount());
     Set<String> iris = new HashSet<>();
     for (TTIriRef entity : iriRefPageable.getResult()) {
       iris.add(entity.getIri());
     }
-    List<EntityReferenceNode> nodes = entityTripleRepository.getEntityReferenceNodes(iris, schemeIris, inactive);
+    List<EntityReferenceNode> nodes = entityRepository.getEntityReferenceNodes(iris, schemeIris, inactive);
     nodes.sort(comparingInt(EntityReferenceNode::getOrderNumber).thenComparing(EntityReferenceNode::getName));
 
     result.setResult(nodes);
@@ -854,27 +562,61 @@ public class EntityService {
     return entityRepository.getIM1SchemeOptions();
   }
 
-  public Set<Concept> getFullyExpandedMembers(String iri, boolean includeLegacy, boolean includeSubset, List<String> schemes) throws QueryException, JsonProcessingException {
-    SetExporter setExporter = new SetExporter();
-    return setExporter.getExpandedSetMembers(iri, true, includeLegacy, includeSubset, schemes);
+
+  public static void isValid(TTEntity entity, String mode) throws TTFilerException, JsonProcessingException {
+    ArrayList<String> errorMessages = new ArrayList();
+    try (CachedObjectMapper om = new CachedObjectMapper()) {
+      if (Boolean.TRUE.equals(!isValidIri(entity))) errorMessages.add("Missing iri.");
+      if ("Create".equals(mode) && iriExists(entity.getIri())) errorMessages.add("Iri already exists.");
+      if ("Update".equals(mode) && !iriExists(entity.getIri())) errorMessages.add("Iri doesn't exists.");
+      if (Boolean.TRUE.equals(!isValidName(entity))) errorMessages.add("Name is invalid.");
+      if (Boolean.TRUE.equals(!isValidType(entity))) errorMessages.add("Types are invalid.");
+      if (Boolean.TRUE.equals(!isValidStatus(entity))) errorMessages.add("Status is invalid");
+      if (Boolean.TRUE.equals(!hasParents(entity))) errorMessages.add("Parents are invalid");
+      if (!errorMessages.isEmpty()) {
+        String errorsAsString = String.join(",", errorMessages);
+        throw new TTFilerException(mode + " entity errors: [" + errorsAsString + "] for entity " + om.writeValueAsString(entity));
+      }
+    }
   }
 
-  public Set<TTIriRef> getSubsets(String iri) {
-    SetExporter setExporter = new SetExporter();
-    return setExporter.getSubsetIrisWithNames(iri);
+  private static Boolean isValidIri(TTEntity entity) {
+    if (null == entity.getIri()) return false;
+    return !"".equals(entity.getIri());
   }
 
-  public List<TTIriRef> getDataModelsFromProperty(String propIri) {
-    return entityRepository.findDataModelsFromProperty(propIri);
-
+  private static Boolean isValidName(TTEntity entity) {
+    if (null == entity.getName()) return false;
+    return !"".equals(entity.getName());
   }
 
-  public List<ConceptContextMap> getConceptContextMaps(String iri) {
-    return entityRepository.getConceptContextMaps(iri);
+  private static Boolean isValidType(TTEntity entity) {
+    if (null == entity.getType()) return false;
+    if (entity.getType().isEmpty()) return false;
+    return entity.getType().getElements().stream().allMatch(TTValue::isIriRef);
   }
 
-  public String checkPropertyType(String iri) {
-    return entityRepository.checkPropertyType(iri);
+  private static Boolean isValidStatus(TTEntity entity) {
+    if (null == entity.getStatus()) return false;
+    return entity.getStatus().isIriRef();
+  }
+
+  private static Boolean hasParents(TTEntity entity) {
+    if (null == entity.get(iri(IM.IS_A)) && null == entity.get(iri(IM.IS_CONTAINED_IN)) && null == entity.get(iri(RDFS.SUBCLASS_OF)) && null == entity.get(iri(IM.IS_SUBSET_OF)))
+      return false;
+    if (null != entity.get(iri(IM.IS_A)) && !entity.get(iri(IM.IS_A)).isEmpty()) {
+      if (!entity.get(iri(IM.IS_A)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    if (null != entity.get(iri(IM.IS_CONTAINED_IN)) && !entity.get(iri(IM.IS_CONTAINED_IN)).isEmpty()) {
+      if (!entity.get(iri(IM.IS_CONTAINED_IN)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    if (null != entity.get(iri(RDFS.SUBCLASS_OF)) && !entity.get(iri(RDFS.SUBCLASS_OF)).isEmpty()) {
+      if (!entity.get(iri(RDFS.SUBCLASS_OF)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    if (null != entity.get(iri(IM.IS_SUBSET_OF)) && !entity.get(iri(IM.IS_SUBSET_OF)).isEmpty()) {
+      if (!entity.get(iri(IM.IS_SUBSET_OF)).getElements().stream().allMatch(TTValue::isIriRef)) return false;
+    }
+    return true;
   }
 }
