--- api/src/test/java/org/endeavourhealth/imapi/logic/service/TestDistillation.java
@@ -1,52 +1,0 @@
-package org.endeavourhealth.imapi.logic.service;
-
-import org.endeavourhealth.imapi.dataaccess.EntityRepository;
-import org.endeavourhealth.imapi.model.tripletree.TTIriRef;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.platform.runner.JUnitPlatform;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-import static org.mockito.Mockito.when;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
-
-@ExtendWith(MockitoExtension.class)
-@RunWith(JUnitPlatform.class)
-public class TestDistillation {
-
-  @InjectMocks
-  EntityService entityService;
-
-  @Mock
-  EntityRepository entityRepository;
-
-  @Test
-  void getDistillation() {
-    List<TTIriRef> conceptList = new ArrayList<>();
-    conceptList.add(new TTIriRef().setIri("http://snomed.info/sct#73211009"));
-    conceptList.add(new TTIriRef().setIri("http://snomed.info/sct#46635009"));
-    conceptList.add(new TTIriRef().setIri("http://snomed.info/sct#44054006"));
-    conceptList.add(new TTIriRef().setIri("http://endhealth.info/im#Q_RegisteredGMS"));
-
-    String iris = "<http://snomed.info/sct#73211009> <http://snomed.info/sct#46635009> <http://snomed.info/sct#44054006> <http://endhealth.info/im#Q_RegisteredGMS>";
-
-    when(entityRepository.getDistillation(iris)).thenReturn(Set.of("http://snomed.info/sct#46635009", "http://snomed.info/sct#44054006"));
-
-    List<TTIriRef> distList = new ArrayList<>();
-    distList.add(new TTIriRef("http://snomed.info/sct#73211009"));
-    distList.add(new TTIriRef("http://endhealth.info/im#Q_RegisteredGMS"));
-
-    List<TTIriRef> actual = entityService.getDistillation(conceptList);
-
-    assertEquals(actual, distList);
-
-  }
-}
